fixed the dll warnings and some other warning
warnings that are coming from dependency some fixed some ignored
DLL, only to export the public functions that is gonna use in editor side
you can export out the entire class as long as there is no  use of std:: classes

+++ GAM300Editor/EditorApp.cpp
@@ -300,7 +300,7 @@ namespace TDS
         (
             runtimePath.c_str(),     // AppDomain base path
             "SampleHost",            // AppDomain friendly name, this can be anything you want really
-            propertyKeys.size(),     // Property count
+            static_cast<int>(propertyKeys.size()),     // Property count
             propertyKeys.data(),     // Property names
             propertyValues.data(),   // Property values
             &hostHandle,             // Host handle
